requirements.txt
tgt==1.4.4
numpy
scipy
loguru
lmdb
torch
pyworld==0.3.0
numba==0.53.0
pyyaml=6.0
torchtext==0.5.0
